ALMOST THERE

+++ api/quotes/delete.php
@@ -21,9 +21,14 @@
     // Set properties/ID to update
     $quote->id = $data->id;
 
+    if (!$quote->exists($quote->id)) {
+        echo json_encode(array("message" => "No Quotes Found"));
+        exit();
+    }
+
     // Delete quote
     if ($quote->delete()) {
-        echo json_encode(array('deleted id' => $quote->id));
+        echo json_encode(array('id' => $quote->id));
     } else {
         echo json_encode(array('message' => 'No Quotes Found'));
     }
